added bcrypt password configuration and sign up module

+++ src/main/kotlin/com/example/log_in_out/authentication/dao/database/token/Token.kt
@@ -10,5 +10,5 @@ data class Token(
         var authenticationToken: String,
 
         @Column(name = "user_id", nullable = false, updatable = false)
-        var userId: Long
+        var userId: Int
 )
